# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Troya e-Ticketing API
  version: 0.1.0
  description: |
    **Troya Modern Biletleme Platformu** — IATA Ticketing Handbook'tan türetilen
    elektronik biletleme (ET) ve EMD modülünün REST kontratı.

    Bu spec **tek doğruluk kaynağıdır** (CLAUDE.md kural 12). Frontend TS tipleri
    `openapi-typescript` ile buradan ÜRETİLİR; elle tip yazılmaz.

    ## Mimari ilkeler (kontrata yansıyan)
    - **Sunum adaptörü (kural 8):** İş kuralı backend otoritesindedir. Client-side
      Zod doğrulaması sadece hızlı geri bildirim içindir.
    - **Para işleminde optimistic UI yok (kural 9):** Issue / exchange / refund /
      void / endorse / EMD / IRROP / PTA komutları sunucu sonucunu bekler.
    - **Idempotency (kural 5):** Para/statü değiştiren her komut bir
      `idempotencyKey` taşır (mock `api.ts` ile uyumlu olarak **request body**
      içinde). Aynı key ile tekrar çağrı = **aynı sonuç, yeni yan etki yok**
      (çift-issue / çift-refund kesinlikle olmaz). Geriye dönük uyum için aynı
      değer `Idempotency-Key` header'ı ile de gönderilebilir.
    - **Coupon Status = explicit FSM (kural 3):** Geçişler `CouponStatus` enum'unda
      tanımlı; izin verilmeyen geçiş `422` problem detayı döndürür. Final statüler
      (`E F G R V P X Z T`) terminaldir.

    ## Domain dili (GLOSSARY)
    Ticket, Coupon, CouponStatus, Validating / Marketing / Operating Carrier,
    Airport Control, EMD, Fare Calculation, TFC (Tax / Fee / Charge), PTA, IRROP / FIM.
...
servers:
...
paths:
...
  /tickets:
    get:
      tags: [Tickets]
      operationId: listTickets
      summary: Bilet özetlerini listele
      description: Tüm biletlerin özet (TicketSummary) listesini döndürür.
      responses:
        "200":
          description: Bilet özet listesi
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TicketSummary" }
    post:
      tags: [Tickets]
      operationId: issueTicket
      summary: Bilet kes (issue)
      description: |
        Yeni bir elektronik bilet (ET) keser. Para işlemidir → optimistic UI yok,
        sunucu sonucu beklenir (kural 9). **Idempotent:** aynı `idempotencyKey`
        ile tekrar = aynı bilet, yeni yan etki yok (kural 5).
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IssueTicketInput" }
      responses:
        "201":
          description: Bilet kesildi (veya idempotent tekrar)
          headers:
            Location:
              description: Kesilen biletin kaynak URL'i
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
  /tickets/{ticketNumber}:
    parameters:
      - $ref: "#/components/parameters/TicketNumberPath"
    get:
      tags: [Tickets]
      operationId: getTicket
      summary: Bilet detayını getir
      description: Tam bilet kaydını (kuponlar, fare, control, history) döndürür.
      responses:
        "200":
          description: Bilet kaydı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "404": { $ref: "#/components/responses/NotFound" }
  /tickets/{ticketNumber}/exchange:
    parameters:
      - $ref: "#/components/parameters/TicketNumberPath"
      - $ref: "#/components/parameters/IdempotencyKeyHeader"
    post:
      tags: [Tickets]
      operationId: exchangeTicket
      summary: Bilet değişimi / reissue (exchange)
      description: |
        Eski biletin açık (`O`) kuponlarını `E` (Exchanged) yapar ve yeni bir bilet
        keser. `adc` = ek tahsilat (Additional Collection); negatif değer residual'i
        ifade eder. Para işlemi → idempotent.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExchangeInput" }
      responses:
        "200":
          description: Eski ve yeni bilet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExchangeResult" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /tickets/{ticketNumber}/refund:
    parameters:
      - $ref: "#/components/parameters/TicketNumberPath"
      - $ref: "#/components/parameters/IdempotencyKeyHeader"
    post:
      tags: [Tickets]
      operationId: refundTicket
      summary: Bilet iadesi (refund)
      description: |
        Seçili açık (`O`) kuponları `R` (Refunded) yapar. Vefat/hastalık waiver'ı
        (Handbook 13.9/13.10, 15.4) iptal cezasını muaf tutabilir. Para işlemi →
        idempotent.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefundInput" }
      responses:
        "200":
          description: Güncellenmiş bilet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /tickets/{ticketNumber}/void:
    parameters:
      - $ref: "#/components/parameters/TicketNumberPath"
      - $ref: "#/components/parameters/IdempotencyKeyHeader"
    post:
      tags: [Tickets]
      operationId: voidTicket
      summary: Bilet iptali (void)
      description: |
        Satış kaydını iptal eder. İnvariant: void için **tüm kuponlar `O`** olmalı;
        değilse `422` döner. Tüm kuponlar `V` (Void) olur. Para işlemi → idempotent.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VoidInput" }
      responses:
        "200":
          description: Void edilmiş bilet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
  /tickets/{ticketNumber}/emds:
    parameters:
      - $ref: "#/components/parameters/TicketNumberPath"
    get:
      tags: [EMD]
      operationId: listEmdsForTicket
      summary: Bilete bağlı EMD'leri listele
      description: Verilen bilete associated (EMD-A) belgeleri döndürür.
      responses:
        "200":
          description: EMD listesi
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Emd" }
    post:
      tags: [EMD]
      operationId: addEmd
      summary: EMD kes (Handbook Ch 5)
      description: |
        Bilet bağlamında bir EMD (Electronic Miscellaneous Document) keser.
        `type` `A` = Associated (bir ET kuponuna bağlı), `S` = Standalone.
        `rfisc` = Reason For Issuance Sub-Code. Para işlemi → idempotent.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddEmdInput" }
      responses:
        "201":
          description: Kesilen EMD
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Emd" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
  /emds:
    get:
      tags: [EMD]
      operationId: listEmds
      summary: EMD'leri listele
      description: |
        Tüm EMD belgelerini döndürür. `associatedTicket` query parametresi ile
        belirli bir bilete filtrelenebilir (mock `listEmdsForTicket` karşılığı).
      parameters:
        - name: associatedTicket
          in: query
          required: false
          description: Sadece bu bilete bağlı (EMD-A) belgeleri döndür
          schema: { type: string }
      responses:
        "200":
          description: EMD listesi
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Emd" }

  # ===================== Dashboard =====================
...
components:
  parameters:
    TicketNumberPath:
      name: ticketNumber
      in: path
      required: true
      description: 13 haneli bilet numarası (3 airline + 9 serial + 1 check digit)
      schema:
        type: string
        pattern: "^\\d{13}$"
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Çift-submit koruması (kural 5). Body içindeki `idempotencyKey` ile aynı
        amaca hizmet eder; geriye dönük uyum / standart uyumu için header olarak da
        kabul edilir. Aynı key ile tekrar = aynı sonuç, yeni yan etki yok.
      schema: { type: string }
  responses:
    BadRequest:
      description: Geçersiz istek
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
    Unauthorized:
      description: Kimlik doğrulama gerekli
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
    NotFound:
      description: Kaynak bulunamadı
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
    Conflict:
      description: Çakışma (örn. idempotency çatışması)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
    UnprocessableEntity:
      description: |
        İş kuralı / FSM ihlali (örn. geçersiz kupon geçişi, void için tüm kuponlar
        `O` değil). Domain hatası burada raporlanır.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
  schemas:
...
    CouponStatus:
      type: string
      description: |
        Coupon Status Indicator — IATA Handbook 1.1.4 (resmî 17 kod).
        Final statüler (`E F G R V P X Z`) terminaldir; bir kupon final statüye
        ulaştıktan sonra değişmez (ARCHITECTURE §4 / kural 3). Not: "T (Paper
        Ticket)" resmî 1.1.4 listelerinde yoktur; kâğıda dönüşümü P/X karşılar.
      enum:
        - O
        - A
        - C
        - L
        - I
        - S
        - U
        - N
        - Y
        - F
        - E
        - G
        - R
        - V
        - P
        - X
        - Z
      x-enum-descriptions:
        O: Open For Use (interim — kupon açık)
        A: Airport Control (interim)
        C: Checked-In (interim)
        L: Lifted / Boarded (interim)
        I: Irregular Operations (interim)
        S: Suspended (interim)
        U: Unavailable (interim)
        N: Notification (interim)
        Y: Refund TFC (interim)
        F: Flown / Used (final)
        E: Exchanged / Reissued (final)
        G: Exchanged / FIM (final)
        R: Refunded (final)
        V: Void (final)
        P: Printed (final)
        X: Print Exchange (final)
        Z: Closed (final)
...
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          description: Tutar (gösterim için number; backend'de minor units olabilir)
        currency:
          type: string
          description: ISO 4217 para birimi kodu
          pattern: "^[A-Z]{3}$"
      additionalProperties: false
    Segment:
      type: object
      description: Yolculuk segmenti (kupon ile 1:1).
      required:
        - origin
        - destination
        - marketingCarrier
        - flightNumber
        - rbd
        - departure
        - arrival
        - fareBasis
        - reservationStatus
      properties:
        origin:
          type: string
          description: Kalkış havalimanı (IATA 3 harf)
          pattern: "^[A-Z]{3}$"
        destination:
          type: string
          description: Varış havalimanı (IATA 3 harf)
          pattern: "^[A-Z]{3}$"
        marketingCarrier:
          type: string
          description: Marketing Carrier (IATA 2 harf)
        operatingCarrier:
          type: string
          description: Operating Carrier (IATA 2 harf)
        flightNumber:
          type: string
        rbd:
          type: string
          description: Booking class (Reservation Booking Designator)
        departure:
          type: string
          format: date-time
        arrival:
          type: string
          format: date-time
        fareBasis:
          type: string
        notValidBefore:
          type: string
          format: date-time
        notValidAfter:
          type: string
          format: date-time
        reservationStatus:
          type: string
          description: 'Rezervasyon statüsü (örn. "HK")'
      additionalProperties: false
    Coupon:
      type: object
      required: [seq, status, segment]
      properties:
        seq:
          type: integer
          minimum: 1
          description: Kupon sıra numarası
        status: { $ref: "#/components/schemas/CouponStatus" }
        segment: { $ref: "#/components/schemas/Segment" }
      additionalProperties: false
...
    FareCalculation:
      type: object
      description: Fare Calculation — baz ücret + TFC toplamı + linear fare calc string.
      required: [baseFare, totalTfc, total, tfcs]
      properties:
        baseFare: { $ref: "#/components/schemas/Money" }
        totalTfc: { $ref: "#/components/schemas/Money" }
        total: { $ref: "#/components/schemas/Money" }
        nuc:
          type: number
          description: Neutral Unit of Construction
        roe:
          type: number
          description: Rate of Exchange
        fareCalcString:
          type: string
          description: Linear fare calculation string
        tfcs:
          type: array
          items: { $ref: "#/components/schemas/TaxFeeCharge" }
      additionalProperties: false
    Passenger:
      type: object
      required: [surname, givenName]
      properties:
        surname: { type: string }
        givenName: { type: string }
        title: { type: string }
        foid:
          type: string
          description: Form of Identification
      additionalProperties: false
    FormOfPayment:
      type: object
      required: [type]
      properties:
        type: { $ref: "#/components/schemas/FormOfPaymentType" }
        detail:
          type: string
          description: Maskelenmiş kart, UATP hesap no vb.
      additionalProperties: false
    ControlAuthority:
      type: object
      description: |
        Airport / Coupon Control — kupon control'ünü aynı anda tek carrier tutar;
        sadece Validating Carrier grant/revoke eder (kural 7). Delegasyonda TTL'li lease.
      required: [holder, isValidatingCarrier]
      properties:
        holder:
          type: string
          description: Control'ü tutan carrier kodu
        isValidatingCarrier:
          type: boolean
        leaseExpiresAt:
          type: string
          format: date-time
          description: Lease bitiş zamanı (sadece delege edildiğinde)
      additionalProperties: false
    LifecycleEvent:
      type: object
      description: Yaşam döngüsü event'i (history endpoint'ini ve timeline'ı besler).
      required: [id, type, occurredAt, actor]
      properties:
        id: { type: string }
        type: { $ref: "#/components/schemas/LifecycleEventType" }
        occurredAt:
          type: string
          format: date-time
        actor:
          type: string
          description: İşlemi yapan carrier / agent
        couponSeq:
          type: integer
        detail:
          type: string
        linkedTicketNumber:
          type: string
          description: Exchange linkage — ilgili diğer bilet numarası
        status:
          $ref: "#/components/schemas/CouponStatus"
      additionalProperties: false
    Ticket:
      type: object
      description: Elektronik bilet (ET) tam kaydı.
      required:
        - ticketNumber
        - passenger
        - validatingCarrier
        - coupons
        - fare
        - formOfPayment
        - control
        - issuedAt
        - history
      properties:
        ticketNumber:
          type: string
          description: 13 hane (3 airline + 9 serial + 1 check digit)
          pattern: "^\\d{13}$"
        passenger: { $ref: "#/components/schemas/Passenger" }
        validatingCarrier:
          type: string
          description: Validating Carrier kodu
        coupons:
          type: array
          items: { $ref: "#/components/schemas/Coupon" }
        fare: { $ref: "#/components/schemas/FareCalculation" }
        formOfPayment: { $ref: "#/components/schemas/FormOfPayment" }
        control: { $ref: "#/components/schemas/ControlAuthority" }
        issuedAt:
          type: string
          format: date-time
        pnr:
          type: string
        endorsement:
          type: string
          description: "Endorsements / Restrictions kutusu (2.19) + ciro notu"
        history:
          type: array
          items: { $ref: "#/components/schemas/LifecycleEvent" }
      additionalProperties: false
    TicketSummary:
      type: object
      required:
        - ticketNumber
        - passengerName
        - route
        - validatingCarrier
        - issuedAt
        - total
        - overallStatus
      properties:
        ticketNumber: { type: string }
        passengerName: { type: string }
        route:
          type: string
          description: 'Örn. "IST → NRT"'
        validatingCarrier: { type: string }
        issuedAt:
          type: string
          format: date-time
        total: { $ref: "#/components/schemas/Money" }
        overallStatus:
          $ref: "#/components/schemas/CouponStatus"
      additionalProperties: false

    # ---------- EMD ----------
...
    Emd:
      type: object
      description: Electronic Miscellaneous Document (Handbook Ch 5).
      required:
        - emdNumber
        - type
        - passenger
        - issuingCarrier
        - issuedAt
        - coupons
        - total
      properties:
        emdNumber:
          type: string
          description: 13 hane
          pattern: "^\\d{13}$"
        type: { $ref: "#/components/schemas/EmdType" }
        passenger: { $ref: "#/components/schemas/Passenger" }
        issuingCarrier: { type: string }
        issuedAt:
          type: string
          format: date-time
        associatedTicket:
          type: string
          description: EMD-A ise bağlı bilet numarası
        associatedCouponSeq:
          type: integer
        coupons:
          type: array
          items: { $ref: "#/components/schemas/EmdCoupon" }
        total: { $ref: "#/components/schemas/Money" }
      additionalProperties: false
...
    IssueTicketInput:
      type: object
      required: [passenger, validatingCarrier, segments, fare, formOfPayment, idempotencyKey]
      properties:
        passenger: { $ref: "#/components/schemas/Passenger" }
        validatingCarrier:
          type: string
          description: Validating Carrier kodu
        pnr:
          type: string
        segments:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/Segment" }
        fare: { $ref: "#/components/schemas/FareCalculation" }
        formOfPayment: { $ref: "#/components/schemas/FormOfPayment" }
        idempotencyKey:
          $ref: "#/components/schemas/IdempotencyKey"
      additionalProperties: false
    VoidInput:
      type: object
      required: [ticketNumber, idempotencyKey]
      properties:
        ticketNumber: { type: string }
        reason: { type: string }
        idempotencyKey: { $ref: "#/components/schemas/IdempotencyKey" }
      additionalProperties: false
    RefundInput:
      type: object
      required: [ticketNumber, couponSeqs, refundAmount, idempotencyKey]
      properties:
        ticketNumber: { type: string }
        couponSeqs:
          type: array
          minItems: 1
          items:
            type: integer
            minimum: 1
        refundAmount: { $ref: "#/components/schemas/Money" }
        residual: { $ref: "#/components/schemas/Money" }
        waiver:
          type: string
          description: Vefat/hastalık muafiyeti (Handbook 13.9/13.10, 15.4)
          enum: [death, illness]
        idempotencyKey: { $ref: "#/components/schemas/IdempotencyKey" }
      additionalProperties: false
    ExchangeInput:
      type: object
      required: [oldTicketNumber, newSegments, adc, idempotencyKey]
      properties:
        oldTicketNumber: { type: string }
        newSegments:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/Segment" }
        adc:
          allOf:
            - $ref: "#/components/schemas/Money"
          description: Additional Collection — ek tahsilat; negatif = residual
        idempotencyKey: { $ref: "#/components/schemas/IdempotencyKey" }
      additionalProperties: false
    ExchangeResult:
      type: object
      required: [oldTicket, newTicket]
      properties:
        oldTicket: { $ref: "#/components/schemas/Ticket" }
        newTicket: { $ref: "#/components/schemas/Ticket" }
      additionalProperties: false
...
    AddEmdInput:
      type: object
      required: [ticketNumber, type, rfisc, description, value, idempotencyKey]
      properties:
        ticketNumber: { type: string }
        couponSeq:
          type: integer
          minimum: 1
          description: EMD-A için bağlanacak kupon sırası
        type: { $ref: "#/components/schemas/EmdType" }
        rfisc:
          type: string
          description: Reason For Issuance Sub-Code
        description: { type: string }
        value: { $ref: "#/components/schemas/Money" }
        idempotencyKey: { $ref: "#/components/schemas/IdempotencyKey" }
      additionalProperties: false
...
    IdempotencyKey:
      type: string
      minLength: 1
      description: |
        Çift-submit koruması (kural 5). Frontend üretir (örn. "idem-<uuid>").
        Aynı key ile tekrar = aynı sonuç, yeni yan etki yok.

    # ---------- Errors ----------
...
  securitySchemes:
    oidc:
      type: openIdConnect
      description: Keycloak OIDC (react-oidc-context ile uçtan uca)
      openIdConnectUrl: /.well-known/openid-configuration
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearer: []